chore: refactors Table
